+++ test/cli/output/test_default.py
@@ -14,8 +14,6 @@ from schemathesis.runner.events import Finished, InternalError
 from schemathesis.runner.models import Check, Request, Status, TestResult, TestResultSet
 from schemathesis.schemas import APIOperation, OperationDefinition
 
-from ...utils import strip_style_win32
-
 
 @pytest.fixture(autouse=True)
 def click_context():
@@ -90,7 +88,6 @@ def test_display_section_name(capsys, title, separator, expected):
     # It should fit into the terminal width
     assert len(click.unstyle(out)) == terminal_width
     # And the section name should be bold
-    assert strip_style_win32(click.style(click.unstyle(out), bold=True)) == out
     assert expected in out
 
 
@@ -106,39 +103,9 @@ def test_handle_initialized(capsys, mocker, execution_context, swagger_20):
     # Then initial title is displayed
     assert " Schemathesis test session starts " in lines[0]
     # And number of collected operations
-    assert strip_style_win32(click.style("Collected API operations: 1", bold=True)) in lines
+    assert "Collected API operations: 1" in out
     # And the output has an empty line in the end
     assert out.endswith("\n")
-
-
-def test_display_statistic(capsys, execution_context, operation, response):
-    # Given multiple successful & failed checks in a single test
-    success = Check("not_a_server_error", Status.SUCCESS, response, 0, operation.Case())
-    failure = Check("not_a_server_error", Status.FAILURE, response, 0, operation.Case())
-    single_test_statistic = TestResult(
-        verbose_name=f"{operation.method} {operation.full_path}",
-        checks=[
-            success,
-            success,
-            success,
-            failure,
-            failure,
-            Check("different_check", Status.SUCCESS, response, 0, operation.Case()),
-        ],
-    )
-    results = TestResultSet(seed=42, results=[single_test_statistic])
-    event = Finished.from_results(results, running_time=1.0)
-    # When test results are displayed
-    default.display_statistic(execution_context, event)
-
-    lines = [line for line in capsys.readouterr().out.split("\n") if line]
-    failed = strip_style_win32(click.style("FAILED", bold=True, fg="red"))
-    passed = strip_style_win32(click.style("PASSED", bold=True, fg="green"))
-    # Then all check results should be properly displayed with relevant colors
-    assert lines[2:4] == [
-        f"    not_a_server_error                    3 / 5 passed          {failed} ",
-        f"    different_check                       1 / 1 passed          {passed} ",
-    ]
 
 
 def test_display_multiple_warnings(capsys, execution_context):
@@ -156,44 +123,11 @@ def test_display_multiple_warnings(capsys, execution_context):
     ]
 
 
-def test_display_statistic_empty(capsys, execution_context, results_set):
-    default.display_statistic(execution_context, Finished.from_results(results_set, running_time=1.23))
-    assert capsys.readouterr().out.split("\n")[2] == strip_style_win32(
-        click.style("No checks were performed.", bold=True)
-    )
-
-
-def test_display_statistic_junitxml(capsys, execution_context, results_set):
-    xml_path = "/tmp/junit.xml"
-    execution_context.junit_xml_file = xml_path
-    default.display_statistic(execution_context, Finished.from_results(results_set, running_time=1.23))
-    assert capsys.readouterr().out.split("\n")[4] == strip_style_win32(
-        click.style("JUnit XML file", bold=True) + click.style(f": {xml_path}")
-    )
-
-
 @pytest.mark.parametrize(
     ("position", "length", "expected"), [(1, 100, "[  1%]"), (20, 100, "[ 20%]"), (100, 100, "[100%]")]
 )
 def test_get_percentage(position, length, expected):
     assert default.get_percentage(position, length) == expected
-
-
-@pytest.mark.parametrize("current_line_length", [0, 20])
-@pytest.mark.parametrize(("operations_processed", "percentage"), [(0, "[  0%]"), (1, "[100%]")])
-def test_display_percentage(
-    capsys, execution_context, after_execution, current_line_length, operations_processed, percentage
-):
-    execution_context.current_line_length = current_line_length
-    execution_context.operations_processed = operations_processed
-    # When percentage is displayed
-    default.display_percentage(execution_context, after_execution)
-    out = capsys.readouterr().out
-    # Then the whole line fits precisely to the terminal width. Note `-1` is padding, that is calculated in a
-    # different place when the line is printed
-    assert len(click.unstyle(out)) + current_line_length - 1 == default.get_terminal_width()
-    # And the percentage displayed as expected in cyan color
-    assert out.strip() == strip_style_win32(click.style(percentage, fg="cyan"))
 
 
 @pytest.mark.parametrize("body", [{}, {"foo": "bar"}, NOT_SET])
@@ -239,26 +173,6 @@ def test_display_single_failure(capsys, execution_context, operation, body, resp
     assert " GET /v1/success " in lines[0]
 
 
-@pytest.mark.parametrize(
-    ("status", "expected_symbol", "color"),
-    [(Status.SUCCESS, ".", "green"), (Status.FAILURE, "F", "red"), (Status.ERROR, "E", "red")],
-)
-def test_handle_after_execution(capsys, execution_context, after_execution, status, expected_symbol, color):
-    # Given AfterExecution even with certain status
-    after_execution.status = status
-    # When this event is handled
-    default.handle_after_execution(execution_context, after_execution)
-
-    assert after_execution.result.verbose_name == "GET /v1/success"
-
-    lines = capsys.readouterr().out.strip().split("\n")
-    symbol, percentage = lines[0].split()
-    # Then the symbol corresponding to the status is displayed with a proper color
-    assert strip_style_win32(click.style(expected_symbol, fg=color)) == symbol
-    # And percentage is displayed in cyan color
-    assert strip_style_win32(click.style("[100%]", fg="cyan")) == percentage
-
-
 def test_after_execution_attributes(execution_context, after_execution):
     # When `handle_after_execution` is executed
     default.handle_after_execution(execution_context, after_execution)
@@ -281,15 +195,3 @@ def test_display_internal_error(capsys, execution_context):
         out = capsys.readouterr().out.strip()
         assert "Traceback (most recent call last):" in out
         assert "ZeroDivisionError: division by zero" in out
-
-
-def test_display_summary(capsys, results_set):
-    # Given the Finished event
-    event = runner.events.Finished.from_results(results=results_set, running_time=1.257)
-    # When `display_summary` is called
-    default.display_summary(event)
-    out = capsys.readouterr().out.strip()
-    # Then number of total tests & total running time should be displayed
-    assert "=== 1 passed in 1.26s ===" in out
-    # And it should be in green & bold style
-    assert strip_style_win32(click.style(click.unstyle(out), fg="green", bold=True)) == out